minor fixes

+++ ST_engine/src/main/game_manager/level/level.hpp
@@ -47,7 +47,7 @@ namespace ST {
         std::vector<ST::text> text_objects{};
         size_t background = 0;
         size_t overlay = 0;
-        int16_t overlay_spriteNum = 1;
+        uint8_t overlay_spriteNum = 1;
         ST::camera Camera = {0, 0, -1, 1920, 0, 1080};
 
         level(const std::string&, message_bus*);
